ci: add GitHub Actions workflow and clean up ruff lint debt
Runs ruff + pytest + repo structure check on push/PR to main. Excludes
notebooks from lint since MANUAL FIRST cells are intentionally
unfinished scaffolding. Fixes pre-existing import-order and redundant
int() cast findings so the gate starts green.

+++ src/pm/fixed_income/bond.py
@@ -1,7 +1,8 @@
 import numpy as np
 
+
 def bond_cashflows(face=100.0, coupon_rate=0.05, years=5.0, frequency=2):
-    n = int(round(years * frequency))
+    n = round(years * frequency)
     coupon = face * coupon_rate / frequency
     flows = np.full(n, coupon, dtype=float)
     flows[-1] += face
